requirements.txt
pandas
streamlit
SQLAlchemy
psycopg2-binary          # Postgres용 드라이버
requests
beautifulsoup4
plotly
ydata-profiling
streamlit-pandas-profiling
geopy
python-dotenv
pandasql